Shift admin config from main app to product app
